Cleanup updateImageDecodePriority
Past changes used a default assignedSize of -1 to see if faces were present for the texture but is unnecessary now, so assignedSize is set to 0 by default and always sent to addTextureStats to avoid branching.

Function local variable names changed to conform with project's naming convention format.

Comments updated and layout adjusted for easier reading.