updateImagesFetchTextures - Fixes for texture references and short loop
Commented out a change by minerjr, fixed a reference in the entries loop and refactored the short loop to use a for loop instead.

The mFetchingTextures loop now double checks there is room in the decode queue because it is now possible to fill the queue on a fast computer with already downloaded files.