Force single thread for updatefetch loop
This entire time I thought it was the threadpool sending too many requests to decode but instead if looks like the compiler was parallelizing the updateFetch loop and causing the issue.

So, hoping a pragma omp single fixes the problem. Seems to be so far but need to test Linux and Mac.