Update Some Texture Debugging Outputs
Adjusted the texture console outputs in lltextureview using the mFetchingTextures size as the number being fetched instead of running a loop over the entire texture to find number actively fetching and moved the decoding queue number closer to the other decode numbers.

Removed a log output in llviewerdisplay that I was using to check the mFetchingTextures was clearing properly and commented out the decode queue full log message (since it can be frequent in some scenarios) in the texture fetch doWork (lltexturefetch).