Avatar Texture Loading Simplification
Removed the updateFetch calls from texture checks to try to avoid a queue deadlock and unnecessary changes to texture properties on checks.

Removed checkTextureLoading in a few processes, limiting it's calling to only the idleUpdate function.

Commented out the reset to 0 in addTextureStats because it seemed to be causing texture fetches to unnecessarily be deleted.

All these changes allow avatars to load from cloud state a lot faster with previous changes and seems to avoid the decode threadpool from being overwhelmed.