Fix Baked Textures on System Avatars
The updateImageDecodePriority runs were setting baked textures without faces to virtual sizes of zero.

Refactored the updateImageDecodePriority function to allow baked textures without faces to not check faces to avoid the zeroing of virtual size and then re-enabled the updateTextures function in the object list's updateApparentAngles paced updates.

Also, commented out the updateApparentAngles in the avatar idle updates, since does not need to be run twice, and adjusted LLVOAvatar::updateTextures() back to a more standard pattern but with importance set to 1 for backed textures.