Narrow MIP change conditions for Octree updates
Adjusting the conditions for the drawable and texture updates from Octree walking for updateDistance so that drawable only added to queue if new MIP is larger than old one and larger than 256 pixel area to reduce number of updates on objects too small to notice changes.

Reductions to drawable and textures will be done in other paced queues.